Move KeyValueStore to kernel
