+++ delta/sdk/src/main/scala/ch/epfl/bluebrain/nexus/delta/sdk/sse/SseConfig.scala
@@ -1,6 +1,6 @@
 package ch.epfl.bluebrain.nexus.delta.sdk.sse
 
-import ch.epfl.bluebrain.nexus.delta.sdk.cache.CacheConfig
+import ch.epfl.bluebrain.nexus.delta.kernel.cache.CacheConfig
 import ch.epfl.bluebrain.nexus.delta.sourcing.config.QueryConfig
 import pureconfig.ConfigReader
 import pureconfig.generic.semiauto.deriveReader
